+++ .gitignore
@@ -1,3 +1,1 @@
-testdata.json
-main
-build/**
+internal/test/testdata.json
