+++ cmd/awsmapping/main.go
@@ -2,11 +2,14 @@ package main
 
 import (
 	"context"
+
 	"github.com/aws/aws-sdk-go-v2/config"
 	"github.com/aws/aws-sdk-go-v2/service/ssoadmin"
-	"go.dfds.cloud/aad-aws-sync/aws"
-	"go.dfds.cloud/aad-aws-sync/util"
+
 	"log"
+
+	"go.dfds.cloud/aad-aws-sync/internal/aws"
+	"go.dfds.cloud/aad-aws-sync/internal/util"
 )
 
 const TIME_FORMAT = "2006-01-02 15:04:05.999999999 -0700 MST"
