+++ cmd/azure2aws/main.go
@@ -2,9 +2,10 @@ package main
 
 import (
 	"fmt"
-	"go.dfds.cloud/aad-aws-sync/azure"
-	"go.dfds.cloud/aad-aws-sync/util"
 	"log"
+
+	"go.dfds.cloud/aad-aws-sync/internal/azure"
+	"go.dfds.cloud/aad-aws-sync/internal/util"
 )
 
 const TIME_FORMAT = "2006-01-02 15:04:05.999999999 -0700 MST"
