+++ internal/aws/model.go
