+++ internal/azure/model.go
