+++ internal/capsvc/model.go
