+++ scripts/get_personal_tokens.sh
